+++ homepage/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @open-tomato/app-homepage
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @open-tomato/ui-portal@0.8.3
+
 ## 0.0.2
 
 ### Patch Changes