+++ packages/shared/theme-default/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @open-tomato/theme-default
 
+## 0.8.3
+
 ## 0.8.2
 
 ## 0.8.1